data-generator/requirements.txt
# ===== PRODUCER NGS - DÉPENDANCES =====
# Version: 1.0.0
# Python: 3.10+

# Kafka Client (Version NG plus maintenue)
kafka-python-ng==2.2.2

# Génération de données
Faker==20.1.0
numpy==1.26.2

# Utilitaires
python-dotenv==1.0.0

# Tests (pour CI/CD)
pytest==7.4.3
pytest-mock==3.12.0


# Instructions pour créer un environnement virtuel Python et installer les dépendances

# 1. Va dans le dossier du générateur
#cd data-generator

# 2. Crée l'environnement virtuel (on l'appelle souvent ".venv" ou "venv")
#python -m venv .venv

# 3. Active-le (Tu dois voir (.venv) apparaître au début de ta ligne de commande)
#.\.venv\Scripts\Activate